1028-2

+++ 42seoul/push_swap/ps_deque_utils.c
@@ -6,13 +6,13 @@
 /*   By: sanghwal <sanghwal@student.42seoul.kr>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/25 17:18:34 by sanghwal          #+#    #+#             */
-/*   Updated: 2022/10/26 15:28:39 by sanghwal         ###   ########seoul.kr  */
+/*   Updated: 2022/10/28 22:09:08 by sanghwal         ###   ########seoul.kr  */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "push_swap.h"
 
-void	dq_push_front(t_deque *stack, int num)
+void	dq_push_front(t_deque *stack, size_t num)
 {
 	size_t	tmp;
 
@@ -27,7 +27,7 @@ void	dq_push_front(t_deque *stack, int num)
 	stack->use_size++;
 }
 
-void	dq_push_rear(t_deque *stack, int num)
+void	dq_push_rear(t_deque *stack, size_t num)
 {
 	size_t	tmp;
 
@@ -43,10 +43,10 @@ void	dq_push_rear(t_deque *stack, int num)
 	return ;
 }
 
-int	dq_pop_front(t_deque *stack)
+size_t	dq_pop_front(t_deque *stack)
 {
 	size_t	tmp;
-	int		num;
+	size_t	num;
 
 	if (stack->front == stack->capacity - 1)
 		tmp = 0;
@@ -59,10 +59,10 @@ int	dq_pop_front(t_deque *stack)
 	return (num);
 }
 
-int	dq_pop_rear(t_deque *stack)
+size_t	dq_pop_rear(t_deque *stack)
 {
 	size_t	tmp;
-	int		num;
+	size_t	num;
 
 	if (stack->rear == 0)
 		tmp = stack->capacity - 1;
